Add code to handle Let's Encrypt registration
This is the first draft of the script that runs when the container
starts, which handles the Certbot first-run.  It does a number of
things:

* It makes /config/logs/letsencrypt, where certbot puts its logs.

* It makes /config/letsencrypt, where certs and keys will be stored.

* It checks for a Let's Encrypt notification email address.

* It checks that you've agreed to the Let's Encrypt ToS.

* It checks for a suitable APP_URL: It must be HTTPS, and have no
  forward-slashes.

Assuming all of that's OK, it runs Certbot registration.  If that fails,
or and environment variables have issues, the script fails.  In those
cases, diagnostics are output to Docker logs.

Since S6 is used here, we use it.  We set ourselves to run once Nginx
setup is complete.