Switch Certbot from Nginx to webroot
Certbot's Nginx plugin can manage the entire renewal process, including
Nginx config changes.  Unfortunately, it makes too many assumptions (for
us, at least) about how Nginx config is laid out.  So, this switches to
the webroot renewal process.

Happily, we can fit this change into the filesystem layout pretty
easily: Bookstack's public-facing web files are located in the container
at /app/www/public, and the existing Nginx config will happily return
the static file if a client requests it.  The Certbot webroot plugin
will also manage the creation (and cleanup) of the
`.well-known/acme-challenge` directory.

The init-certbot-config script gets a lot of changes.  `certbot run` is
changed to use the webroot plugin, and provide a path.  Once the cert is
issued, the Nginx SSL config is then changed to point to the new cert,
key, and chanin.  We also turn on OCSP Stalping, since we have all the
bit necessary to do it.

The renewal Periodic job is also updated, to specify webroot renewal,
although it might not be necessary to specify that, since we specified
it when we first got the cert.

Finally, the Docker file is updated to no longer bring in the `certbot-nginx`
package.