init-certbot-config: Fail on colon in APP_URL
I just realized that Let's Encrypt will always use port 80 when it's
reaching out to us in order to check for our response to the http-01
challenge.  See RFC 8555 Section 8.3: The URL template is defined as
`"http://{domain}/.well-known/acme-challenge/{token}"`; this forces
Let's Encrypt to use port 80.

So, that means the Booksack site has to be able to respond on port 80.
Therefore, we might as well just prohibit port numbers in the APP_URL.

This change implements that restriction.  As a result, the APP_URL
becomes much easier to parse!