Move to Bookstack 23.12.3
See https://github.com/BookStackApp/BookStack/releases/tag/v23.12 for
the list of improvements for 23.12.  The .1 and .2 releases are smaller
bugfixes, and .3 brings a security fix related to PDF generation.

This also brings in two SAML-related improvements
(BookStackApp/BookStack#4713 and BookStackApp/BookStack#4706).  Thanks
again ssddanbrown for them!  And thanks also to the LinuxServer folks
for the packaging.